+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v1.0.4
  * monaca: Added `Monaca.getLatestNews()` to fetch latest news and status on known issues from Monaca Cloud.
  * localkit: Added `options.clientId` to `Localkit.startProject()` method. Used to start a project on one device instead of sending the start signal to all connected devices.
  * localkit: Added `Localkit.startInspector()` to start an inspector when the computer is connected to Localkit.
+ * monoaca: Added `Monaca.isMonacaProject()` to check if a directory is a Monaca project.
 
 v1.0.3
 ----